chore: storybook에 ThemeProvider와 전역 스타일링 적용 (#3)
